Designathon/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
python-docx
PyPDF2
pdfplumber
transformers==4.36.2
torch==2.1.2 